requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.19
celery==3.1.17
Django==1.8
django-admin-bootstrapped==2.5.0
django-appconf==1.0.1
django-audit-log==0.7.0
django-bower==5.0.4
django-compressor==1.5
django-filter==0.10.0
django-taggit==0.13.0
django-taggit-serializer==0.1.1
django-websocket-redis==0.4.4
djangorestframework==3.1.1
djangorestframework-jwt==1.6.0
dpkt==1.8.6
gevent==1.0.2
greenlet==0.4.6
influxdb==2.9.2
kombu==3.0.24
msgpack-python==0.4.6
psycopg2==2.6.1
PyJWT==1.3.0
pymongo==3.0.1
python-dateutil==2.4.2
pytz==2015.2
redis==2.10.3
requests==2.8.1
scapy==2.3.1
six==1.9.0
socket.io-emitter==0.1.3
uWSGI==2.0.11.1
